fix: Rewrote the extension using own logic
