+++ test/testCPreprocessorExtension.py
@@ -1,6 +1,6 @@
 import unittest
 from mock import Mock, patch
-from lizard_ext.lizardcpre import LizardExtension as CPreprocessor
+from lizard_ext.lizardcpreprocess import LizardExtension as CPreprocessor
 from .testHelpers import get_cpp_function_list_with_extension
 from lizard_languages.code_reader import CodeReader
 generate_tokens = CodeReader.generate_tokens
